Module Bls12_381.Ff_sig

module type BASE = sig ... end

Base module signature for a finite field

module type PRIME = sig ... end

Module type for prime field of the form GF(p) where p is prime. The order of GF(p) is p

module type PRIME_WITH_ROOT_OF_UNITY = sig ... end

Module type for prime field with additional functions to manipulate roots of unity